1、客户端远程调用，目前支持随机方式的负载均衡。 2、实现注解注入远程服务。 3、实现生产者消费者两端服务上传下达 4、消费者，注册中心，生产者三端联调测试已通过。

+++ SourceCode/loxfdubbo/dubbo/registry/src/main/java/org/loxf/registry/bean/Server.java
@@ -34,7 +34,7 @@ public class Server extends BaseBean {
 	/**
 	 * 失效时间毫秒（注册中心timeout时长过后，删除server信息）
 	 */
-	private int timeout;
+	private long timeout;
 	
 	public Server(String serverAddr, int serverPort){
 		this.serverAddr = serverAddr;
@@ -110,7 +110,7 @@ public class Server extends BaseBean {
 	/**
 	 * @return the timeout
 	 */
-	public int getTimeout() {
+	public long getTimeout() {
 		return timeout;
 	}
 
@@ -118,7 +118,7 @@ public class Server extends BaseBean {
 	 * @param timeout
 	 *            the timeout to set
 	 */
-	public void setTimeout(int timeout) {
+	public void setTimeout(long timeout) {
 		this.timeout = timeout;
 	}
 }
